+++ backend/libraries/types/src/community_summary.rs
@@ -18,7 +18,7 @@ pub struct CommunityCanisterCommunitySummary {
     pub description: String,
     pub avatar_id: Option<u128>,
     pub banner_id: Option<u128>,
-    pub is_public: bool,
+    pub is_public: Option<bool>,
     pub member_count: u32,
     pub permissions: CommunityPermissions,
     pub frozen: Option<FrozenGroupInfo>,
@@ -27,20 +27,20 @@ pub struct CommunityCanisterCommunitySummary {
     pub latest_event_index: EventIndex,
     pub channels: Vec<CommunityCanisterChannelSummary>,
     pub membership: Option<CommunityMembership>,
-    pub user_groups: Vec<UserGroupSummary>,
+    pub user_groups: Option<Vec<UserGroupSummary>>,
     pub is_invited: Option<bool>,
     pub metrics: ChatMetrics,
-    pub verified: bool,
+    pub verified: Option<bool>,
 }
 
 #[ts_export]
 #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
 pub struct CommunityMembership {
     pub joined: TimestampMillis,
-    pub role: CommunityRole,
-    pub rules_accepted: bool,
+    pub role: Option<CommunityRole>,
+    pub rules_accepted: Option<bool>,
     pub display_name: Option<String>,
-    pub lapsed: bool,
+    pub lapsed: Option<bool>,
 }
 
 #[ts_export]
@@ -50,25 +50,25 @@ pub struct CommunityCanisterCommunitySummaryUpdates {
     pub last_updated: TimestampMillis,
     pub name: Option<String>,
     pub description: Option<String>,
-    #[ts(as = "crate::OptionUpdateU128")]
-    pub avatar_id: OptionUpdate<u128>,
-    #[ts(as = "crate::OptionUpdateU128")]
-    pub banner_id: OptionUpdate<u128>,
+    #[ts(as = "Option<crate::OptionUpdateU128>")]
+    pub avatar_id: Option<OptionUpdate<u128>>,
+    #[ts(as = "Option<crate::OptionUpdateU128>")]
+    pub banner_id: Option<OptionUpdate<u128>>,
     pub is_public: Option<bool>,
     pub member_count: Option<u32>,
     pub permissions: Option<CommunityPermissions>,
-    #[ts(as = "crate::OptionUpdateFrozenGroupInfo")]
-    pub frozen: OptionUpdate<FrozenGroupInfo>,
-    #[ts(as = "crate::OptionUpdateAccessGateConfig")]
-    pub gate_config: OptionUpdate<AccessGateConfig>,
+    #[ts(as = "Option<crate::OptionUpdateFrozenGroupInfo>")]
+    pub frozen: Option<OptionUpdate<FrozenGroupInfo>>,
+    #[ts(as = "Option<crate::OptionUpdateAccessGateConfig>")]
+    pub gate_config: Option<OptionUpdate<AccessGateConfig>>,
     pub primary_language: Option<String>,
     pub latest_event_index: Option<EventIndex>,
-    pub channels_added: Vec<CommunityCanisterChannelSummary>,
-    pub channels_updated: Vec<CommunityCanisterChannelSummaryUpdates>,
-    pub channels_removed: Vec<ChannelId>,
+    pub channels_added: Option<Vec<CommunityCanisterChannelSummary>>,
+    pub channels_updated: Option<Vec<CommunityCanisterChannelSummaryUpdates>>,
+    pub channels_removed: Option<Vec<ChannelId>>,
     pub membership: Option<CommunityMembershipUpdates>,
-    pub user_groups: Vec<UserGroupSummary>,
-    pub user_groups_deleted: Vec<u32>,
+    pub user_groups: Option<Vec<UserGroupSummary>>,
+    pub user_groups_deleted: Option<Vec<u32>>,
     pub metrics: Option<ChatMetrics>,
     pub verified: Option<bool>,
 }
@@ -78,7 +78,7 @@ pub struct CommunityCanisterCommunitySummaryUpdates {
 pub struct CommunityMembershipUpdates {
     pub role: Option<CommunityRole>,
     pub rules_accepted: Option<bool>,
-    #[ts(as = "crate::OptionUpdateString")]
-    pub display_name: OptionUpdate<String>,
+    #[ts(as = "Option<crate::OptionUpdateString>")]
+    pub display_name: Option<OptionUpdate<String>>,
     pub lapsed: Option<bool>,
 }